Initial Session Rework
Merged Detector Controler and MoodControler into new SessionController
Moved Detector views and Mood views into new Session view folder
Changed _Layout to display only Sessions
