+++ MoodDetectorWebApp/Controllers/SessionController.cs
@@ -5,20 +5,74 @@ using Model;
 using Model.Entity;
 using MoodDetectorWebApp.Models;
 using System;
+using Newtonsoft.Json;
 
 namespace MoodDetectorWebApp.Controllers
 {
-    [Authorize(Roles = "teacher")]
-    public class MoodController : Controller
+    public class SessionController : Controller
     {
         IMoodService _moodService;
         IUserService _userService;
 
-        public MoodController(IMoodService moodService, IUserService userService)
+        public SessionController(IMoodService moodService, IUserService userService)
         {
             _moodService = moodService;
             _userService = userService;
         }
+
+        public ActionResult Session()
+        {
+            return View();
+        }
+
+        //NEW SESSION
+
+        // GET: Detector
+        [HttpGet]
+        public ActionResult NewSession()
+        {
+            return View(new NewSessionModel());
+        }
+
+        // POST: Detector
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult NewSession(NewSessionModel model)
+        {
+            // Randamas dabartinis vartotojas is prisijungimo metu sukurto cookio.
+            User currentUser = _userService.GetUser(System.Web.HttpContext.Current.User.Identity.Name);
+
+            if (ModelState.IsValid)
+            {
+                SessionInfo sessionInfo = new SessionInfo()
+                {
+                    User = currentUser,
+                    Class = model.Class,
+                    Comments = model.Comments,
+                    Subject = model.Subject,
+                    DateTime = DateTime.Now,
+                    MessageSeen = 0,
+                };
+
+                int detectionId = _moodService.AddSession(sessionInfo);
+                DetectionStartViewModel detectionStartViewModel = new DetectionStartViewModel()
+                {
+                    DetectionId = detectionId,
+                };
+
+                return View("~/Views/Detector/Start.cshtml", detectionStartViewModel);
+            }
+
+            return View();
+        }
+
+        // POST: Detector/PostMoods/
+        public void PostMoods(int detectionId, string moods)
+        {
+            _moodService.AddMood(detectionId, JsonConvert.DeserializeObject<MoodCollection>(moods));
+        }
+
+        //EXISTING SESSIONS
 
         public ActionResult MoodIndex()
         {
@@ -29,7 +83,7 @@ namespace MoodDetectorWebApp.Controllers
         {
             return View("SessionIndex");
         }
-
+        
         public ActionResult SessionsList()
         {
             int userId = _userService.GetUser(System.Web.HttpContext.Current.User.Identity.Name).Id;
@@ -54,7 +108,7 @@ namespace MoodDetectorWebApp.Controllers
         public ActionResult GetMoodList(int id)
         {
             List<Mood> moods = _moodService.GetMoodsByDate(new User() { Id = id });
-            
+
             return View("MoodList", moods);
         }
