+++ MoodDetectorWebApp/Views/Session/MoodIndex.cshtml
