+++ MoodDetectorWebApp/Views/Session/MoodList.cshtml
