+++ MoodDetectorWebApp/Views/Session/NewSession.cshtml
