+++ MoodDetectorWebApp/Views/Session/Session.cshtml
@@ -1,0 +1,9 @@
+﻿
+
+<div class="row mt-5 mb-3">
+    <div class="col-lg-11 col-sm-10 col-9">@Html.ActionLink("Create new Session", "NewSession", "Session", null, new { @class = "btn btn-secondary btn-large" })</div>
+</div>
+
+<div class="row  mb-3">
+    <div class="col-lg-11 col-sm-10 col-9">@Html.ActionLink("View Sessions", "SessionsList", "Session", null, new { @class = "btn btn-secondary btn-large" })</div>
+</div>
