+++ MoodDetectorWebApp/Views/Session/SessionIndex.cshtml
