+++ MoodDetectorWebApp/Views/Session/SessionsList.cshtml
