+++ MoodDetectorWebApp/Views/Session/Start.cshtml
